Day 9 of Advent
Day 9 of Advent

+++ DayFive/DayFive.cs
@@ -2,21 +2,21 @@
 namespace DayFive;
 public class DayFive {
 
-    public static uint run(){
+    public static long run(){
         return doPart1();
     }
 
-    public static uint doPart1(){
-        (List<uint>, Finder) items = Parser.parse("../DayFive/TextExamples/Messages.txt");
+    public static long doPart1(){
+        (List<long>, Finder) items = Parser.parse("../DayFive/TextExamples/Messages.txt");
         return getAllseeds(items);
   
     }
 
-    private static uint getAllseeds((List<uint>, Finder) items){
-        uint smallest = uint.MaxValue;
+    private static long getAllseeds((List<long>, Finder) items){
+        long smallest = long.MaxValue;
         for(int i = 0; i < items.Item1.Count;i += 2){
-            for(uint l = items.Item1[i]; l < items.Item1[i] + items.Item1[i + 1];l++){
-                uint location = items.Item2.findLocationFromSeed(l);
+            for(long l = items.Item1[i]; l < items.Item1[i] + items.Item1[i + 1];l++){
+                long location = items.Item2.findLocationFromSeed(l);
                 if(location < smallest){
                     smallest = location;
                 }
