+++ DayFive/Parser.cs
@@ -1,9 +1,9 @@
 public class Parser{
 
-    public static (List<uint>,Finder) parse(string FileName){
+    public static (List<long>,Finder) parse(string FileName){
         int place = 0;
         Finder finder = new Finder();
-        List<uint> seeds = new List<uint>();
+        List<long> seeds = new List<long>();
         try{
             StreamReader sr = new StreamReader(FileName);
             string? line = sr.ReadLine();
@@ -32,19 +32,19 @@ public class Parser{
 
     private static Range parseRange(string line){
         string[] parts = line.Split(" ");
-        uint[] partNums = new uint[]{uint.Parse(parts[0]),uint.Parse(parts[1]),uint.Parse(parts[2])};
-        (uint,uint) range = (partNums[1],partNums[1] + partNums[2] - 1);
-        uint diff = partNums[1] - partNums[0];
+        long[] partNums = new long[]{long.Parse(parts[0]),long.Parse(parts[1]),long.Parse(parts[2])};
+        (long,long) range = (partNums[1],partNums[1] + partNums[2] - 1);
+        long diff = partNums[1] - partNums[0];
         return new Range(range, diff);
     }
 
-    private static List<uint> parseSeeds(string? line){
+    private static List<long> parseSeeds(string? line){
         if(line == null) line = ""; //To shut up warnings
         string[] parts = line.Split(":");
         string[] seedsS = parts[1].Trim().Split(" ");
-        List<uint> seeds = new List<uint>();
+        List<long> seeds = new List<long>();
         foreach(string seed in seedsS){
-            seeds.Add(uint.Parse(seed));
+            seeds.Add(long.Parse(seed));
         }
         return seeds;
     }
